replace perl in example date parsing
Time::Piece was released with perl v5.8.9 but may not be installed on
restricted systems.  Hardcoding the number of days is less flexible but
"only" relies on a GNU date supporting -d.

Furthermore, git only supports date formatting since version 2.6.0,
so prefer an equivalent cut similar to earlier year selection.